pyproject.toml
[tool.poetry]
name = "teste-frexco"
version = "0.1.0"
description = ""
authors = ["SilasAlvesTi <silas.ifce.ti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
httpie = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
